A program which runs 4 different types of sort algorithms and keeps track of their load
This program will generate large arrays of random numbers. The numnber generated is 10^x. ie, when the user inputs 'h, 6', the program will run heap sort on 10^6 numbers, then 10^5, and so on down to one. It then displays the amount of operations that were done on the data. 

It has the ability to run four sort algorithms, quicksort, merge sort, heap sort and selection sort. For each sort, it will keep track of the comparisons and moves of data.